requirements.txt
python-telegram-bot[rate-limiter]==22.1
docker==7.1.0